Revert "Merge branch 'main' into Oscar"
This reverts commit fb1b702d551bed4e8a8bf92503ef9c9489fa334a.

+++ GameJam 2024 rigtige/Assets/Scripts/Player/Sword.cs
@@ -3,7 +3,7 @@ using System.Collections;
 using System.Collections.Generic;
 public class Sword : MonoBehaviour
 {
-    int damage;
+    float damage;
     List<Enemy> enemysList = new List<Enemy>();
     public float angle;
